telegram-bot

+++ internal/usecase/admin_service.go
@@ -20,22 +20,36 @@ func (s *AdminService) RegisterUser(iin, password, fullName, role string) error 
 		return errors.New("IIN must be 12 digits")
 	}
 
+	if role != "admin" && role != "doctor" {
+		return errors.New("role must be either 'admin' or 'doctor'")
+	}
+
 	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
 	if err != nil {
 		return err
 	}
+
 	return s.repo.RegisterUser(iin, string(hashedPassword), fullName, role)
 }
 
 func (s *AdminService) BlockUser(iin string) error {
+	if len(iin) != 12 {
+		return errors.New("IIN must be 12 digits")
+	}
 	return s.repo.BlockUser(iin)
 }
 
 func (s *AdminService) UnblockUser(iin string) error {
+	if len(iin) != 12 {
+		return errors.New("IIN must be 12 digits")
+	}
 	return s.repo.UnblockUser(iin)
 }
 
 func (s *AdminService) DeleteUser(iin string) error {
+	if len(iin) != 12 {
+		return errors.New("IIN must be 12 digits")
+	}
 	return s.repo.DeleteUser(iin)
 }
